changed classes

+++ src/main/CountingExpression.java
@@ -6,9 +6,10 @@ import java.util.LinkedList;
  * Created by Nelly
  */
 public class CountingExpression extends ExpressionParser {
+
     //The method calculates the value of the expression, already converted to polish notation
-    protected Integer counting(String value) {
-        LinkedList<Integer> numbers = new LinkedList<>();
+    protected Long counting(String value) {
+        LinkedList<Long> numbers = new LinkedList<>();
         LinkedList<Character> opers = new LinkedList<>();
         for (int i = 0; i < value.length(); i++) {
             char c = value.charAt(i);
@@ -29,7 +30,7 @@ public class CountingExpression extends ExpressionParser {
                 while (i < value.length() && Character.isDigit(value.charAt(i)))
                     operand += value.charAt(i++);
                 --i;
-                numbers.add(Integer.parseInt(operand));
+                numbers.add(Long.parseLong(operand));
             }
         }
         while (!opers.isEmpty())
@@ -37,9 +38,9 @@ public class CountingExpression extends ExpressionParser {
         return numbers.get(0);
     }
 
-    private void processOperator(LinkedList<Integer> numbers, char operator) {
-        int oneElement = numbers.removeLast();
-        int twoElement = numbers.removeLast();
+    private void processOperator(LinkedList<Long> numbers, char operator) {
+        long oneElement = numbers.removeLast();
+        long twoElement = numbers.removeLast();
         switch (operator) {
             case '+':
                 numbers.add(twoElement + oneElement);
